Simplify config file script in sdk
